requirements/base.txt
Django==1.7.3
dj-database-url==0.3.0
django-cache-url==0.8.0
django-configurations==0.8
gunicorn==19.1.1
psycopg2==2.5.4
